Refactor sparkey and use package object for import
